+++ satella/instrumentation/metrics/labeled.py
@@ -1,5 +1,6 @@
 import typing as tp
 
+from .metric_types import CallableMetric
 from .metric_types.measurable_mixin import MeasurableMixin
 from .metric_types.base import Metric, MetricLevel
 
@@ -21,12 +22,18 @@ class LabeledMetric(Metric, MeasurableMixin):
 
     >>> b.runtime(1)
     """
-    __slots__ = ('metric_to_wrap', 'labels')
+    __slots__ = ('metric_to_wrap', 'labels', 'callable')
 
     def __init__(self, metric_to_wrap, **labels):
         self.metric_to_wrap = metric_to_wrap
         self.labels = labels
+        self.callable = labels.pop('callable', lambda: 0)
+
+        if isinstance(metric_to_wrap, CallableMetric):
+            metric_to_wrap.register_labeled_metric(self)
 
     def handle(self, level: tp.Union[int, MetricLevel], *args, **kwargs) -> None:
         kwargs.update(self.labels)
         self.metric_to_wrap.handle(level, *args, **kwargs)
+
+